reordenamiento y download file and start game funcs
